requirements.txt
speedtest-cli
psutil
tqdm